// src/app/core/models/delegaciones.ts
export class modelDelegaciones{
    rnc: string
    firmanteAutorizado: boolean;
    solicitanteAutorizado: boolean;
    aprobadorComercial: boolean;
    administrador: boolean;
    estado: string;
    identificacion:string;
    fechaRegistro: string;
    fechaActualizacion: string;
    AmbienteID: number;
    CanalID: number;
}